Remove TxVerifier
