+++ examples/demo-stf/src/app.rs
@@ -1,7 +1,6 @@
 #[cfg(feature = "native")]
 use crate::runner_config::Config;
 use crate::runtime::Runtime;
-use crate::tx_verifier_impl::DemoAppTxVerifier;
 use sov_default_stf::AppTemplate;
 pub use sov_default_stf::Batch;
 use sov_default_stf::SequencerOutcome;
@@ -42,7 +41,7 @@ use sov_modules_macros::expose_rpc;
 #[cfg(feature = "native")]
 pub type NativeAppRunner<Vm> = DemoAppRunner<DefaultContext, Vm>;
 
-pub type DemoApp<C, Vm> = AppTemplate<C, DemoAppTxVerifier<C>, Runtime<C>, Vm>;
+pub type DemoApp<C, Vm> = AppTemplate<C, Runtime<C>, Vm>;
 
 /// Batch receipt type used by the demo app. We export this type so that it's easily accessible to the full node.
 pub type DemoBatchReceipt = SequencerOutcome;
@@ -59,8 +58,7 @@ impl<Vm: Zkvm> StateTransitionRunner<ProverConfig, Vm> for DemoAppRunner<Default
         let runtime = Runtime::new();
         let storage = ProverStorage::with_config(runtime_config.storage)
             .expect("Failed to open prover storage");
-        let tx_verifier = DemoAppTxVerifier::new();
-        let app = AppTemplate::new(storage, runtime, tx_verifier);
+        let app = AppTemplate::new(storage, runtime);
         Self(app)
     }
 
@@ -80,13 +78,8 @@ impl<Vm: Zkvm> StateTransitionRunner<ZkConfig, Vm> for DemoAppRunner<ZkDefaultCo
     fn new(runtime_config: Self::RuntimeConfig) -> Self {
         let runtime = Runtime::new();
         let storage = ZkStorage::with_config(runtime_config).expect("Failed to open zk storage");
-        let tx_verifier = DemoAppTxVerifier::new();
-        let app: AppTemplate<
-            ZkDefaultContext,
-            DemoAppTxVerifier<ZkDefaultContext>,
-            Runtime<ZkDefaultContext>,
-            Vm,
-        > = AppTemplate::new(storage, runtime, tx_verifier);
+        let app: AppTemplate<ZkDefaultContext, Runtime<ZkDefaultContext>, Vm> =
+            AppTemplate::new(storage, runtime);
         Self(app)
     }
